terrible hack to support closed captions
The CC decoder is stolen from VLC. The hard part about supporting CCs is
that in ffmpeg, CCs are returned by the video decoder. This is due to
the terrible fact that CCs are stored in video packets, and also need to
be reordered like video frames. So we have a few hacks in demux.c and
vd_lavc.c to pretend that CCs work like normal subtitles.

Also, since there is no way to probe CCs without decoding video, CC
tracks are added only after playback starts. They must be selected at
runtime.

The CCs are for now output as plaintext; this might be not so good.